Small changes of data types

+++ save_filedata.h
@@ -41,9 +41,9 @@ struct saveCharState
 {
     saveCharState() : id(1), state(1), itemID(0), mountType(0), mountID(0), health(1) {}
     //! ID of playable character
-    int id;
+    unsigned long id;
     //! Curent ID of playable character's state
-    unsigned int state;
+    unsigned long state;
     //! Current item ID in the item slot (SMBX64-only)
     unsigned long itemID;
     //! Mounted vehicle type (SMBX64-only)
@@ -79,7 +79,7 @@ struct GamesaveData
     long worldPosX; //!< Last world map position X
     long worldPosY; //!< Last world map position Y
 
-    long last_hub_warp; //!< Last entered/exited warp Array-ID on the HUB-based episodes.
+    unsigned long last_hub_warp; //!< Last entered/exited warp Array-ID on the HUB-based episodes.
 
     unsigned int musicID; //!< Current world music ID
     PGESTRING musicFile;    //!< Current world music file (custom music)
@@ -87,7 +87,7 @@ struct GamesaveData
     bool gameCompleted;   //!< Is episode was completed in last time
 
     PGEVECTOR<saveCharState > characterStates;
-    PGELIST<int > currentCharacter;
+    PGELIST<unsigned long > currentCharacter;
 
     //Visible state of world map items
     PGEVECTOR<visibleItem > visibleLevels;
@@ -97,4 +97,3 @@ struct GamesaveData
 };
 
 #endif // SAVE_FILEDATA_H
-
